+++ Pod/Classes/ViewControllers/ARArtistComponentViewController.m
@@ -18,9 +18,4 @@
   return self;
 }
 
-- (UIStatusBarStyle)preferredStatusBarStyle
-{
-    return UIStatusBarStyleLightContent;
-}
-
 @end