+++ Pod/Classes/ViewControllers/ARMySellingProfileViewController.m
@@ -1,0 +1,10 @@
+#import "ARMySellingProfileViewController.h"
+
+@implementation ARMySellingProfileViewController
+
+- (instancetype)init
+{
+    return [super initWithEmission:nil moduleName:@"MyProfileEdit" initialProperties:nil];
+}
+
+@end