+++ src/lib/Components/Inbox/Conversations/Messages.tsx
@@ -7,6 +7,8 @@ import Message from "./Message"
 import ArtworkPreview from "./Preview/ArtworkPreview"
 import ShowPreview from "./Preview/ShowPreview"
 
+const isPad = Dimensions.get("window").width > 700
+
 interface Props {
   conversation?: RelayProps["me"]["conversation"]
   relay?: RelayPaginationProp
@@ -101,12 +103,20 @@ export class Messages extends React.Component<Props, State> {
     })
     const refreshControl = <RefreshControl refreshing={this.state.reloadingData} onRefresh={this.reload.bind(this)} />
 
+    const messagesStyles = isPad
+      ? {
+          width: 708,
+          alignSelf: "center",
+        }
+      : {}
+
     return (
       <FlatList
         inverted={!this.state.shouldStickFirstMessageToTop}
         data={this.state.shouldStickFirstMessageToTop ? messages.reverse() : messages}
         renderItem={this.renderMessage.bind(this)}
         keyExtractor={({ __id }) => __id}
+        keyboardShouldPersistTaps="always"
         onEndReached={this.loadMore.bind(this)}
         onEndReachedThreshold={0.2}
         onContentSizeChange={(width, height) => {
@@ -125,6 +135,7 @@ export class Messages extends React.Component<Props, State> {
           })
         }}
         refreshControl={refreshControl}
+        style={messagesStyles}
         ListFooterComponent={<ActivityIndicator animating={this.state.fetchingMoreData} hidesWhenStopped />}
       />
     )