+++ src/lib/Components/Inbox/Conversations/Preview/ArtworkPreview.tsx
@@ -1,18 +1,20 @@
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 
-import { TouchableHighlight } from "react-native"
+import { Text, TouchableHighlight } from "react-native"
 
 import { PreviewText as P, Subtitle } from "../../Typography"
 
+import { Schema, Track, track as _track } from "../../../../utils/track"
+
 import OpaqueImageView from "lib/Components/OpaqueImageView"
-import colors from "lib/data/colors"
-import fonts from "lib/data/fonts"
+import { Colors } from "lib/data/colors"
+import { Fonts } from "lib/data/fonts"
 import styled from "styled-components/native"
 
 const Container = styled.View`
   border-width: 1;
-  border-color: ${colors["gray-regular"]};
+  border-color: ${Colors.GrayRegular};
   flex-direction: row;
 `
 
@@ -24,13 +26,13 @@ const VerticalLayout = styled.View`
 const Image = styled(OpaqueImageView)`
   margin-top: 12;
   margin-left: 12;
+  margin-right: 12;
   margin-bottom: 12;
   width: 80;
   height: 55;
 `
 
 const TextContainer = styled(VerticalLayout)`
-  margin-left: 25;
   align-self: center;
 `
 
@@ -38,31 +40,38 @@ const SerifText = styled(P)`
   font-size: 14;
 `
 
+const Date = styled.Text`font-family: ${Fonts.GaramondRegular};`
+
 const TitleAndDate = styled.View`
   margin-top: 3;
+  margin-right: 12;
   flex-direction: row;
+  justify-content: flex-start;
 `
-
-const Title = styled.Text`
-  font-family: ${fonts["garamond-italic"]};
-  flex: 3;
-  font-size: 14;
-`
-
-const Date = styled(SerifText)`
-  flex: 1;
-`
-
 interface Props extends RelayProps {
   onSelected?: () => void
 }
 
+const track: Track<Props> = _track
+
+@track()
 export class ArtworkPreview extends React.Component<Props, any> {
+  @track((props, state) => ({
+    action_type: Schema.ActionTypes.Tap,
+    action_name: Schema.ActionNames.ConversationAttachmentArtwork,
+    owner_type: Schema.OwnerEntityTypes.Artwork,
+    owner_id: props.artwork._id,
+    owner_slug: props.artwork.id,
+  }))
+  attachmentSelected() {
+    this.props.onSelected()
+  }
+
   render() {
     const artwork = this.props.artwork
 
     return (
-      <TouchableHighlight underlayColor={colors["gray-light"]} onPress={this.props.onSelected}>
+      <TouchableHighlight underlayColor={Colors.GrayLight} onPress={() => this.attachmentSelected()}>
         <Container>
           <Image imageURL={artwork.image.url} />
           <TextContainer>
@@ -70,10 +79,11 @@ export class ArtworkPreview extends React.Component<Props, any> {
               {artwork.artist_names}
             </SerifText>
             <TitleAndDate>
-              <Title numberOfLines={1}>
-                {artwork.title}
-              </Title>
-              {!!artwork.date && <Date>{`, ${artwork.date}`}</Date>}
+              {/* Nested Text components are necessary for the correct behaviour on both short and long titles + dates */}
+              <Subtitle numberOfLines={1} ellipsizeMode={"middle"}>
+                {`${artwork.title}`}
+                <Date>{`, ${artwork.date}`}</Date>
+              </Subtitle>
             </TitleAndDate>
           </TextContainer>
         </Container>
@@ -86,6 +96,8 @@ export default createFragmentContainer(
   ArtworkPreview,
   graphql`
     fragment ArtworkPreview_artwork on Artwork {
+      id
+      _id
       title
       artist_names
       date
@@ -98,6 +110,8 @@ export default createFragmentContainer(
 
 interface RelayProps {
   artwork: {
+    id: string | null
+    _id: string | null
     title: string | null
     artist_names: string | null
     date: string | null