+++ src/lib/Components/Inbox/Conversations/Preview/__tests__/ShowPreview-tests.tsx
@@ -16,6 +16,8 @@ it("renders correctly for a fair booth", () => {
 
 const show = {
   name: "Catty Show",
+  id: "slugID",
+  _id: "mongoID",
   cover_image: {
     url: "https://d32dm0rphc51dk.cloudfront.net/bJ9I_vJX9ksaKFJAkOAIKg/normalized.jpg",
   },